+++ packages/winnow/src/test/filter.js
@@ -13,6 +13,13 @@ test('With a where option', (t) => {
   run('trees', options, 12105, t)
 })
 
+test('With a field with a space', (t) => {
+  const options = {
+    where: '"total precip" > 1'
+  }
+  run('snow', options, 4, t)
+})
+
 test('With esri json', (t) => {
   const options = {
     where: "Genus like '%Quercus%'",