+++ packages/winnow/CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Support esri-style spatial predicates
 * Support esri-style aggregations
 
+### Fixed
+* Support fields like `total precip`
+
 ## [1.0.1] - 03-29-2016
 ### Fixed
 * Handle esri-style envelopes without a spatialReference